+++ src/main/java/edu/ftn/iss/eventplanner/repositories/CategoryRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface CategoryRepository extends JpaRepository<SolutionCategory, Long> {
+public interface CategoryRepository extends JpaRepository<SolutionCategory, Integer> {
     Optional<SolutionCategory> findByName(@Nonnull String name);
 }
